Updated default icon for styled select form component
